refactor: reconstruct FloatingIconButton component in accordance with atomic design
