+++ src/components/atoms/button/floating-icon-button/PostButton.tsx
@@ -1,0 +1,12 @@
+import { FaPencilAlt } from 'react-icons/fa';
+import BaseFloatingIconButton from './BaseFloatingIconButton';
+
+type Props = {
+  onClick: () => void;
+};
+
+export default function PostButton({ onClick }: Props) {
+  return (
+    <BaseFloatingIconButton className="right-3 bottom-3" onClick={onClick} icon={<FaPencilAlt />} />
+  );
+}
